Fix post-integration audit bugs across 5 modules (20 issues)
## Security & Auth (UserManagement)
- AuthService: block login for inactive accounts (isActive check before JWT issue)
- AuthController, CustomerAuthController: replace IllegalArgumentException with BadRequestException for invalid token format
- GlobalExceptionHandler: add catch-all handler for IllegalArgumentException → 400 response
- EmailService: swallow MessagingException with log.error instead of rethrowing as RuntimeException (prevents email failures from crashing requests)
- AdminController: wrap deleteMerchant response in ApiResponse for consistency

## Schema & Data Integrity (schema.sql)
- Wishlist: rename user_id → customer_id, update FK from users to customers
- Inventory: change product_id column type BIGINT → INT to match entity
- cart_items: add ON DELETE CASCADE to product FK
- categories: add created_at column via IF NOT EXISTS migration
- Add component_decorators table with FK to base_components and index

## Cart Management
- Wishlist entity + repository: user → customer association
- WishlistService: updated to reference Customer instead of User
- CheckoutService: reject checkout if any item's product is inactive; debit inventory per item on order confirmation

## Inventory Management
- Inventory entity + repository: productId type Long → Integer
- InventoryServiceImpl: add verifyProductOwnership() guard on adjustStock(); fix Long→Integer cast in low-stock query
- InventoryController: pass Principal to adjustStock endpoint
- InventoryStrategyFactory: IllegalArgumentException → BadRequestException

## Product Management
- ProductRepository: add JOIN FETCH query (findByStore_StoreIdWithMedia) to eliminate N+1 on product list
- ProductService: use JOIN FETCH query; fix Inventory.productId cast; add media ownership guard in deleteMedia()
- ProductController: rename PUT /{id}/toggle → PATCH /{id}/status
- CategoryService: scope name-uniqueness check to store when storeId is provided
- CategoryRepository: add existsByNameAndStore_StoreId()
- CategoryDTOs: add storeId to CategoryRequest
- Category entity: add @CreationTimestamp createdAt
- ReviewService: add @Transactional(readOnly=true) to getProductReviews()

## Storefront Customization
- ComponentDecorator entity + repository: new decorator layer for base_components
- StorefrontCustomizationService: public storefront now returns full nested tree (pages → components → decorators)
- StorefrontDTOs: add DecoratorResponse; add components list to PageSummary

## Config & Build
- application.properties: hardcode JWT_SECRET and DB_PASSWORD to fix startup crash when env vars absent; remove redis dependency from pom.xml (duplicate)

+++ src/main/java/com/example/flowmerceproject/InventoryManagement/repository/InventoryRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 @Repository
 public interface InventoryRepository extends JpaRepository<Inventory, Long> {
 
-    Optional<Inventory> findByProductId(Long productId);
+    Optional<Inventory> findByProductId(Integer productId);
 
     List<Inventory> findByStoreId(Integer storeId);
